+++ shared/i18n/locales/en_US/translation.json
@@ -284,28 +284,20 @@
   "Results": "Results",
   "No results for {{query}}": "No results for {{query}}",
   "Manage": "Manage",
+  "All members": "All members",
+  "Everyone in the workspace": "Everyone in the workspace",
   "Invite": "Invite",
   "{{ userName }} was added to the collection": "{{ userName }} was added to the collection",
   "{{ count }} people added to the collection": "{{ count }} people added to the collection",
   "{{ count }} people added to the collection_plural": "{{ count }} people added to the collection",
   "{{ count }} people and {{ count2 }} groups added to the collection": "{{ count }} people and {{ count2 }} groups added to the collection",
   "{{ count }} people and {{ count2 }} groups added to the collection_plural": "{{ count }} people and {{ count2 }} groups added to the collection",
-  "All members": "All members",
-  "Everyone in the workspace": "Everyone in the workspace",
   "Add": "Add",
   "Add or invite": "Add or invite",
   "Viewer": "Viewer",
   "Editor": "Editor",
   "Suggestions for invitation": "Suggestions for invitation",
   "No matches": "No matches",
-  "{{ userName }} was removed from the document": "{{ userName }} was removed from the document",
-  "Could not remove user": "Could not remove user",
-  "Permissions for {{ userName }} updated": "Permissions for {{ userName }} updated",
-  "Could not update user": "Could not update user",
-  "Has access through <2>parent</2>": "Has access through <2>parent</2>",
-  "Suspended": "Suspended",
-  "Invited": "Invited",
-  "Leave": "Leave",
   "Can view": "Can view",
   "Everyone in the collection": "Everyone in the collection",
   "You have full access": "You have full access",
@@ -314,6 +306,14 @@
   "Other workspace members may have access": "Other workspace members may have access",
   "This document may be shared with more workspace members through a parent document or collection you do not have access to": "This document may be shared with more workspace members through a parent document or collection you do not have access to",
   "Access inherited from collection": "Access inherited from collection",
+  "{{ userName }} was removed from the document": "{{ userName }} was removed from the document",
+  "Could not remove user": "Could not remove user",
+  "Permissions for {{ userName }} updated": "Permissions for {{ userName }} updated",
+  "Could not update user": "Could not update user",
+  "Has access through <2>parent</2>": "Has access through <2>parent</2>",
+  "Suspended": "Suspended",
+  "Invited": "Invited",
+  "Leave": "Leave",
   "Only lowercase letters, digits and dashes allowed": "Only lowercase letters, digits and dashes allowed",
   "Sorry, this link has already been used": "Sorry, this link has already been used",
   "Public link copied to clipboard": "Public link copied to clipboard",
